Create Maximum_path_weighted_tree.py
m
m

What is the Weighted Tree?
It is a tree in which every edge will be assigned with a number. It is also called edge distance between two connected nodes.

Explanation
The class node is used to save the data for each tree node.

The class variable intVal is used to save the edge value from its parent. For example, the value of intVal for node ‘C’ is 3, for node ‘E’ is 4 and for node ‘A’ (root node) is 0 (zero).

A listPtr is the Python list object. It will save the addresses of all the child nodes.

To find the max path in the weighted tree, we are using the recursion mechanism.